requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
